DomainServices factory method
