+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/metrics/RepositoryGauges.scala
@@ -3,7 +3,7 @@ package pl.touk.nussknacker.ui.metrics
 import io.dropwizard.metrics5.{CachedGauge, Gauge, MetricName, MetricRegistry}
 import pl.touk.nussknacker.engine.api.deployment.ScenarioActionName
 import pl.touk.nussknacker.ui.process.ScenarioQuery
-import pl.touk.nussknacker.ui.process.repository.DBFetchingProcessRepository
+import pl.touk.nussknacker.ui.process.repository.FetchingProcessRepository
 import pl.touk.nussknacker.ui.security.api.{LoggedUser, NussknackerInternalUser}
 
 import java.time.Duration
@@ -15,7 +15,7 @@ import scala.concurrent.duration.DurationInt
 class RepositoryGauges(
     metricRegistry: MetricRegistry,
     repositoryGaugesCacheDuration: Duration,
-    processRepository: DBFetchingProcessRepository[Future]
+    processRepository: FetchingProcessRepository[Future]
 ) {
 
   private val awaitTime = 5 seconds
